Commiting custom nebula and @config changes. Note aspace.cnf needs to be in the game directory next to mush.cnf, and local.c needs to be slightly modified for this to work.
Namely the hashinit of &local_options needs to be increased.
##  hashinit(&local_options, 20);

Due to the nature of the @config changes, the hash table previously used for consoles is now deprecated. As well as the old console_(name) dbrefs.

Everything is stored in the aspace_config structure, including the spacetime cochrane value.

The #define CONSOLE_NAME (Dbref) in space.h are still valid unless over written by the dbrefs in aspace.cnf

We also now announce when entering/exiting a nebula.